Add files via upload
The primary objective of this project is to develop a content-based recommendation system that simplifies cosmetic product selection based on ingredient similarity, with a specific focus on moisturizers for dry skin. The project aims to use data science techniques to create an interactive visualization that enables users to understand product formulations and make safer, more informed skincare decisions.